# docs/swagger/gen/user/v1/swagger.yaml
components:
    schemas:
...
        request.CreateSpotRequest:
            properties:
                description:
                    description: 説明
                    type: string
                latitude:
                    description: 緯度
                    type: number
                longitude:
                    description: 経度
                    type: number
                name:
                    description: スポット名
                    type: string
                spotTypeId:
                    description: スポット種別ID
                    type: string
                thumbnailUrl:
                    description: サムネイルURL
                    type: string
            type: object
...
        request.UpdateSpotRequest:
            properties:
                description:
                    description: 説明
                    type: string
                latitude:
                    description: 緯度
                    type: number
                longitude:
                    description: 経度
                    type: number
                name:
                    description: スポット名
                    type: string
                spotTypeId:
                    description: スポット種別ID
                    type: string
                thumbnailUrl:
                    description: サムネイルURL
                    type: string
            type: object
...
        response.Spot:
            description: スポット情報
            properties:
                createdAt:
                    description: 登録日時
                    type: integer
                description:
                    description: 説明
                    type: string
                id:
                    description: スポットID
                    type: string
                latitude:
                    description: 座標情報:緯度
                    type: number
                longitude:
                    description: 座標情報:経度
                    type: number
                name:
                    description: スポット名
                    type: string
                spotTypeId:
                    description: スポット種別ID
                    type: string
                thumbnailUrl:
                    description: サムネイル画像URL
                    type: string
                updatedAt:
                    description: 更新日時
                    type: integer
                userId:
                    description: 投稿者のユーザーID
                    type: string
                userName:
                    description: 投稿者名
                    type: string
                userThumbnailUrl:
                    description: 投稿者のサムネイルURL
                    type: string
                userType:
                    description: 投稿者の種別
                    type: integer
            type: object
        response.SpotResponse:
            properties:
                spot:
                    $ref: '#/components/schemas/response.Spot'
                spotType:
                    $ref: '#/components/schemas/response.SpotType'
            type: object
        response.SpotType:
            description: スポット種別情報
            properties:
                id:
                    description: 体験種別ID
                    type: string
                name:
                    description: 体験種別名
                    type: string
            type: object
        response.SpotTypesResponse:
            properties:
                spotTypes:
                    description: 体験種別一覧
                    items:
                        $ref: '#/components/schemas/response.SpotType'
                    type: array
                    uniqueItems: false
                total:
                    description: 体験種別合計数
                    type: integer
            type: object
        response.SpotsResponse:
            properties:
                spotTypes:
                    description: スポット種別一覧
                    items:
                        $ref: '#/components/schemas/response.SpotType'
                    type: array
                    uniqueItems: false
                spots:
                    description: スポット一覧
                    items:
                        $ref: '#/components/schemas/response.Spot'
                    type: array
                    uniqueItems: false
            type: object
...
        util.ErrorResponse:
            properties:
                detail:
                    description: エラー詳細
                    type: string
                message:
                    description: エラー概要
                    type: string
                status:
                    description: ステータスコード
                    type: integer
            type: object
    securitySchemes:
        bearerauth:
            bearerFormat: JWT
            scheme: bearer
            type: http
        cookieauth:
            in: cookie
            name: session_id
            type: apiKey
...
info:
    description: 購入者向けのふるマルAPIです。
    title: ふるマル API - 購入者向け
    version: ""
openapi: 3.1.0
paths:
...
    /spot-types:
        get:
            description: スポットタイプの一覧を取得します。
            parameters:
                - description: 取得件数
                  in: query
                  name: limit
                  schema:
                    default: 20
                    type: integer
                - description: 取得開始位置
                  in: query
                  name: offset
                  schema:
                    default: 0
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.SpotTypesResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
            summary: スポットタイプ一覧取得
            tags:
                - SpotType
    /spots:
        get:
            description: 指定された位置情報周辺のスポット一覧を取得します。
            parameters:
                - description: 緯度
                  in: query
                  name: latitude
                  required: true
                  schema:
                    type: number
                - description: 経度
                  in: query
                  name: longitude
                  required: true
                  schema:
                    type: number
                - description: 検索半径（km）
                  in: query
                  name: radius
                  schema:
                    default: 20
                    type: integer
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.SpotsResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
            summary: スポット一覧取得
            tags:
                - Spot
        post:
            description: 新しいスポットを登録します。
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.CreateSpotRequest'
                description: スポット情報
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.SpotResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
            security:
                - bearerauth: []
            summary: スポット登録
            tags:
                - Spot
    /spots/{spotId}:
        delete:
            description: 指定されたIDのスポットを削除します。
            parameters:
                - description: スポットID
                  in: path
                  name: spotId
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: 削除成功
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: アクセス権限がありません
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: スポットが見つかりません
            security:
                - bearerauth: []
            summary: スポット削除
            tags:
                - Spot
        get:
            description: 指定されたIDのスポット詳細を取得します。
            parameters:
                - description: スポットID
                  in: path
                  name: spotId
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/response.SpotResponse'
                    description: OK
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: スポットが見つかりません
            summary: スポット詳細取得
            tags:
                - Spot
        patch:
            description: 指定されたIDのスポット情報を更新します。
            parameters:
                - description: スポットID
                  in: path
                  name: spotId
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/request.UpdateSpotRequest'
                description: 更新するスポット情報
                required: true
            responses:
                "204":
                    description: 更新成功
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: バリデーションエラー
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: 認証エラー
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: アクセス権限がありません
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/util.ErrorResponse'
                    description: スポットが見つかりません
            security:
                - bearerauth: []
            summary: スポット更新
            tags:
                - Spot
...
servers:
    - description: 検証環境
      url: https://api.furumaru-stg.and-period.work
    - description: 本番環境
      url: https://api.furumaru.and-period.co.jp